+++ BALSAMIC/assets/scripts/cap_base_quality_in_bam.py
@@ -1,12 +1,13 @@
 import click
 import pysam
+import numpy as np
 
 
 @click.command()
 @click.argument("input_bam", type=click.Path(exists=True))
 @click.argument("output_bam", type=click.Path())
 @click.option(
-    "--max_quality", default=70, type=int, help="Maximum quality value to cap to."
+    "--max-quality", default=70, type=int, help="Maximum quality value to cap to."
 )
 def cap_base_qualities(input_bam: str, output_bam: str, max_quality: int):
     """
@@ -18,17 +19,15 @@ def cap_base_qualities(input_bam: str, output_bam: str, max_quality: int):
         max_quality (int): Maximum quality value to cap to.
     """
     # Open input BAM file for reading
-    with pysam.AlignmentFile(input_bam, "rb") as in_bam:
-        # Open output BAM file for writing
-        with pysam.AlignmentFile(output_bam, "wb", header=in_bam.header) as out_bam:
-            # Iterate over reads in the input BAM file
-            for read in in_bam:
-                # Cap base qualities to the maximum value
-                capped_qualities = [min(q, max_quality) for q in read.query_qualities]
-                # Update the base qualities in the read
-                read.query_qualities = capped_qualities
-                # Write the modified read to the output BAM file
-                out_bam.write(read)
+    samfile = pysam.AlignmentFile(input_bam, "rb")
+    out_bam = pysam.AlignmentFile(output_bam, "wb", header=samfile.header)
+    for read in samfile.fetch():
+        qualities = np.array(read.query_qualities)
+        capped_qualities = np.minimum(qualities, max_quality)
+        # Update the base qualities in the read
+        read.query_qualities = capped_qualities.tolist()
+        # Write the modified read to the output BAM file
+        out_bam.write(read)
 
 
 if __name__ == "__main__":