+++ BALSAMIC/constants/cache.py
@@ -52,6 +52,7 @@ class DockerContainers(StrEnum):
     GATK: str = "gatk"
     HTSLIB: str = "htslib"
     MULTIQC: str = "multiqc"
+    MSISENSORPRO: str = "msisensorpro"
     PURECN: str = "purecn"
     PYTHON_3: str = "varcall_py3"
     PYTHON_27: str = "varcall_py27"