+++ BALSAMIC/constants/workflow_params.py
@@ -104,6 +104,9 @@ VCF_DICT = {
 SLEEP_BEFORE_START = 600
 
 WORKFLOW_PARAMS = {
+    "bam_post_processing": {
+        "manta_max_base_quality": 70,
+    },
     "common": {
         "pcr_model": "NONE",
         "min_mapq": "20",
@@ -140,9 +143,21 @@ WORKFLOW_PARAMS = {
             ]
         ),
     },
+    "insert_size_metrics": {
+        "min_read_ratio": 0.01,
+    },
     "manta": {
         "wgs_settings": "",
         "tga_settings": "--exome",
+    },
+    "mosdepth": {
+        "mapq": 20,
+        "samflag": 1796,
+        "quantize": "0:1:50:150:",
+    },
+    "sentieon_wgs_metrics": {
+        "min_base_qual": 10,
+        "cov_threshold": [50, 100, 150, 200, 250],
     },
     "vep": {
         "vep_filters": "--compress_output bgzip --vcf --everything --hgvsg --allow_non_variant --dont_skip --buffer_size 30000 --max_sv_size 249250621 --format vcf --offline --variant_class --merged --cache --verbose --force_overwrite"